requirements.txt
asgiref==3.3.1
Django==3.1.6
django-multiselectfield==0.1.12
djangorestframework==3.12.2
Pillow==8.1.0
pytils==0.3
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
